+++ By_ToolsExample/Assets/3rdBy/ByFramework/UI/Editor/UIAutoCreate/UIControlAutoCreate.cs
@@ -62,7 +62,7 @@ namespace _3rdBy.ByFramework.UI.Editor.UIAutoCreate
                 }
             }
 
-            Debug.Log("创建成功: " + filePath);
+            Debug.Log("[UI自动生成器] Control 代码创建成功：" + filePath);
             AssetDatabase.Refresh();
         }
     }